+++ molecularnodes/__init__.py
@@ -1,6 +1,8 @@
 from .ui.addon import register, unregister
 from .entities import fetch, parse
+from . import session
 from . import color, blender
+from . import ui
 
 try:
     from .scene import Canvas
