+++ molecularnodes/ui/addon.py
@@ -15,12 +15,14 @@ import bpy
 from bpy.app.handlers import frame_change_pre, load_post, save_post
 from bpy.props import PointerProperty, CollectionProperty
 
-from . import pref, props
+from . import pref, props, panel, ops, node_menu
 from ..handlers import update_entities
 from ..utils import add_current_module_to_path
-from .. import session, ui
+from .. import session
 
-all_classes = ui.CLASSES + props.CLASSES + pref.CLASSES + session.CLASSES
+all_classes = (
+    panel.CLASSES + ops.CLASSES + props.CLASSES + pref.CLASSES + session.CLASSES
+)
 
 
 def _test_register():
@@ -41,9 +43,9 @@ def register():
             # print(e)
             pass
     add_current_module_to_path()
-    bpy.types.NODE_MT_add.append(ui.node_menu.MN_add_node_menu)
-    bpy.types.VIEW3D_MT_object_context_menu.prepend(ui.panel.pt_object_context)
-    bpy.types.NODE_MT_context_menu.prepend(ui.panel.change_style_node_menu)
+    bpy.types.NODE_MT_add.append(node_menu.MN_add_node_menu)
+    bpy.types.VIEW3D_MT_object_context_menu.prepend(panel.pt_object_context)
+    bpy.types.NODE_MT_context_menu.prepend(panel.change_style_node_menu)
 
     save_post.append(session._pickle)
     load_post.append(session._load)
@@ -66,9 +68,9 @@ def unregister():
             # print(e)
             pass
 
-    bpy.types.NODE_MT_add.remove(ui.node_menu.MN_add_node_menu)
-    bpy.types.VIEW3D_MT_object_context_menu.remove(ui.panel.pt_object_context)
-    bpy.types.NODE_MT_context_menu.remove(ui.panel.change_style_node_menu)
+    bpy.types.NODE_MT_add.remove(node_menu.MN_add_node_menu)
+    bpy.types.VIEW3D_MT_object_context_menu.remove(panel.pt_object_context)
+    bpy.types.NODE_MT_context_menu.remove(panel.change_style_node_menu)
 
     save_post.remove(session._pickle)
     load_post.remove(session._load)
