+++ tests/conftest.py
@@ -6,7 +6,7 @@ from .utils import NumpySnapshotExtension
 import molecularnodes as mn
 
 
-mn.addon._test_register()
+mn.ui.addon._test_register()
 
 
 DATA_DIR = join(dirname(realpath(__file__)), "data")
